DEV: include force option in link, unlink and diffuse commands

+++ syndot/commands.py
@@ -57,19 +57,27 @@ def link(args: Namespace) -> None:
                 if os.path.exists(source_target_path):
                     print(f"Skipping already linked {source_target_path}")
                 else:
-                    print(f"{source_target_path} is a symlink to the missing {destination_target_path}")
+                    print(f"{source_target_path} is a symlink to {destination_target_path}, which is missing")
             else:
                 if os.path.exists(destination_target_path):
-                    question = f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, " \
-                               f"not to {destination_target_path} \nForce link to {destination_target_path}"
-                    force_link = utils.prompt_question(question = question, default = 'n')
-                    if force_link:
+                    if args.force:
                         utils.remove(path = source_target_path)
                         diffuse_dotfile(source_target_path = source_target_path,
                                         destination_target_path = destination_target_path)
                         if args.backup:
                             backup_path = utils.generate_backup_path(path = source_target_path)
                             utils.copy(source = destination_target_path, destination = backup_path)
+                    else:
+                        question = f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, " \
+                                   f"not to {destination_target_path} \nForce link to {destination_target_path}"
+                        force_link = utils.prompt_question(question = question, default = 'n')
+                        if force_link:
+                            utils.remove(path = source_target_path)
+                            diffuse_dotfile(source_target_path = source_target_path,
+                                            destination_target_path = destination_target_path)
+                            if args.backup:
+                                backup_path = utils.generate_backup_path(path = source_target_path)
+                                utils.copy(source = destination_target_path, destination = backup_path)
                 else:
                     print(f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, "
                           f"not to {destination_target_path}, which does not exists")
@@ -92,14 +100,20 @@ def unlink(args: Namespace) -> None:
                                        source = source,
                                        destination = destination)
                     else:
-                        question = f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, " \
-                                   f"not to {destination_target_path} \nForce unlink of {destination_target_path}"
-                        force_unlink = utils.prompt_question(question = question, default = 'n')
-                        if force_unlink:
+                        if args.force:
                             unlink_dotfile(source_target_path = source_target_path,
                                            destination_target_path = destination_target_path,
                                            source = source,
                                            destination = destination)
+                        else:
+                            question = f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, " \
+                                       f"not to {destination_target_path} \nForce unlink of {destination_target_path}"
+                            force_unlink = utils.prompt_question(question = question, default = 'n')
+                            if force_unlink:
+                                unlink_dotfile(source_target_path = source_target_path,
+                                               destination_target_path = destination_target_path,
+                                               source = source,
+                                               destination = destination)
                 else:
                     if not os.path.exists(source_target_path):
                         unlink_dotfile(source_target_path = source_target_path,
@@ -107,13 +121,20 @@ def unlink(args: Namespace) -> None:
                                        source = source,
                                        destination = destination)
                     else:
-                        question = f"{source_target_path} is not a symlink\nForce unlink of {destination_target_path}"
-                        force_unlink = utils.prompt_question(question = question, default = 'n')
-                        if force_unlink:
+                        if args.force:
                             unlink_dotfile(source_target_path = source_target_path,
                                            destination_target_path = destination_target_path,
                                            source = source,
                                            destination = destination)
+                        else:
+                            question = f"{source_target_path} is not a symlink\n" \
+                                       f"Force unlink of {destination_target_path}"
+                            force_unlink = utils.prompt_question(question = question, default = 'n')
+                            if force_unlink:
+                                unlink_dotfile(source_target_path = source_target_path,
+                                               destination_target_path = destination_target_path,
+                                               source = source,
+                                               destination = destination)
             else:
                 print(f"Skipping missing {destination_target_path}")
         else:
@@ -135,23 +156,34 @@ def diffuse(args: Namespace) -> None:
                         diffuse_dotfile(source_target_path = source_target_path,
                                         destination_target_path = destination_target_path)
                     else:
-                        question = f"{source_target_path} already exists\nForce diffuse of {destination_target_path}"
-                        force_diffuse = utils.prompt_question(question = question, default = 'n')
-                        if force_diffuse:
+                        if args.force:
                             utils.remove(source_target_path)
                             diffuse_dotfile(source_target_path = source_target_path,
                                             destination_target_path = destination_target_path)
+                        else:
+                            question = f"{source_target_path} already exists\n"\
+                                       f"Force diffuse of {destination_target_path}"
+                            force_diffuse = utils.prompt_question(question = question, default = 'n')
+                            if force_diffuse:
+                                utils.remove(source_target_path)
+                                diffuse_dotfile(source_target_path = source_target_path,
+                                                destination_target_path = destination_target_path)
                 else:
                     if os.readlink(source_target_path) == destination_target_path:
                         print(f"Skipping {destination_target_path} because already linked by {source_target_path}")
                     else:
-                        question = f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, " \
-                                   f"not to {destination_target_path} \nForce diffuse of {destination_target_path}"
-                        force_diffuse = utils.prompt_question(question = question, default = 'n')
-                        if force_diffuse:
+                        if args.force:
                             utils.remove(source_target_path)
                             diffuse_dotfile(source_target_path = source_target_path,
                                             destination_target_path = destination_target_path)
+                        else:
+                            question = f"{source_target_path} is a symlink to {os.readlink(source_target_path)}, " \
+                                       f"not to {destination_target_path} \nForce diffuse of {destination_target_path}"
+                            force_diffuse = utils.prompt_question(question = question, default = 'n')
+                            if force_diffuse:
+                                utils.remove(source_target_path)
+                                diffuse_dotfile(source_target_path = source_target_path,
+                                                destination_target_path = destination_target_path)
             else:
                 print(f"Skipping missing {destination_target_path}")
         else:
